Support for node globals in production by using mendel-browser-pack
