+++ packages/mendel-middleware/index.js
@@ -3,7 +3,7 @@
    See the accompanying LICENSE file for terms. */
 
 var pathToRegexp = require('path-to-regexp');
-var bpack = require('browser-pack');
+var mbpack = require('mendel-browser-pack');
 var MendelTrees = require('mendel-core');
 var MendelLoader = require('mendel-loader');
 var debug = require('debug')('mendel:middleware');
@@ -104,19 +104,18 @@ function MendelMiddleware(opts) {
             return next();
         }
 
-        var pack = bpack({raw: true, hasExports: true});
         var decodedResults = trees.findTreeForHash(params.bundle, params.hash);
         if (!decodedResults || decodedResults.error) {
             return bundleError(res, 404, decodedResults && decodedResults.error, params);
         }
 
-        var modules = indexedDeps(decodedResults.deps.filter(Boolean));
+        var modules = decodedResults.deps.filter(Boolean);
 
         if (!modules.length) {
             // Something wrong, modules shouldn't be zero
             return bundleError(res, 500, {
                 code: 'EMPTYBUNDLE',
-                message: 'Tree contents are empty'
+                message: 'Tree contents are empty',
             }, params);
         }
 
@@ -126,115 +125,9 @@ function MendelMiddleware(opts) {
             'Cache-Control': 'public, max-age=31536000',
         });
 
+        var pack = mbpack(modules);
         pack.pipe(res);
-        for (var i = 0; i < modules.length; i++) {
-            pack.write(modules[i]);
-        }
-        pack.end();
     };
-}
-
-/*
-Here is a piece of hard to read JavaScript.
-This compresses the bundle by renaming all dependency indexes from file paths
-to a numbered index.
-
-Here is a sample transformation:
-[
-  {
-    "entry": true,
-    "id": "/User/me/projects/site/src/main.js",
-    "deps": {
-      "./colors.js": "/User/me/projects/site/src/colors.js",
-      "./shared.js": "/User/me/projects/site/src/shared.js"
-    }
-  },
-  {
-    "id": "/User/me/projects/site/src/colors.js",
-    "deps": {
-      "external-lib": false
-    }
-  },
-  {
-    "expose": "shared",
-    "id": "/User/me/projects/site/src/shared.js",
-    "deps": {}
-  }
-]
-
-Should become:
-[
-  {
-    "entry": true,
-    "id": 1,
-    "deps": {
-      "./colors.js": 2,
-      "./shared.js": "shared"
-    }
-  },
-  {
-    "id": 2,
-    "deps": {
-      "external-lib": false
-    }
-  },
-  {
-    "expose": "shared",
-    "id": "shared",
-    "deps": {}
-  }
-]
-
-*/
-
-function indexedDeps(mods) {
-    // the index can't be ever 0 because 0 is false for browserify
-    var newModIndex = [0];
-
-    // indexes are created first, because deps can come unordered
-    mods.forEach(function(mod){
-        if (!mod.expose) newModIndex.push(mod.id);
-    });
-
-    // create a new array of modified modules
-    return mods.map(function(oldMod) {
-        return Object.keys(oldMod).reduce(function(newMod, prop) {
-
-            if (prop === 'deps') { // deps needs to be reindexed
-                newMod.deps = Object.keys(oldMod.deps).reduce(
-                    function(newDeps, name) {
-                        var id = oldMod.deps[name];
-                        var index = newModIndex.indexOf(id);
-                        if (index > -1) {
-                            newDeps[name] = index;
-                        } else {
-                            // deps not indexed are exposed or external
-                            newDeps[name] = id;
-                        }
-                        return newDeps;
-                    },
-                {});
-            }
-
-
-            else if(prop === 'id') { // id needs to be reindexed
-                var index = newModIndex.indexOf(oldMod.id);
-                if (index > -1) {
-                    newMod.id = index;
-                } else {
-                    // unless it is entry or exposed
-                    newMod.id = oldMod.expose || oldMod.id;
-                }
-            }
-
-            else {
-                // for all other props we just copy over
-                newMod[prop] = oldMod[prop];
-            }
-
-            return newMod;
-        }, {});
-    });
 }
 
 /******
