+++ packages/mendel-middleware/package.json
@@ -21,8 +21,8 @@
     "url": "https://github.com/yahoo/mendel"
   },
   "dependencies": {
-    "browser-pack": "^6.0.1",
     "debug": "^2.6.3",
+    "mendel-browser-pack": "3.0.0",
     "mendel-core": "^3.0.3",
     "mendel-loader": "^3.0.0",
     "path-to-regexp": "^1.2.1"
