+++ packages/mendel-outlet-browser-pack/package.json
@@ -13,7 +13,7 @@
     "url": "https://github.com/yahoo/mendel"
   },
   "dependencies": {
-    "browser-pack": "^6.0.2",
-    "debug": "^2.6.0"
+    "debug": "^2.6.0",
+    "mendel-browser-pack": "3.0.0"
   }
 }
