+++ packages/mendel-outlet-manifest/src/index.js
@@ -52,6 +52,7 @@ module.exports = class ManifestOutlet {
 
         const data = {
             id: item.normalizedId,
+            normalizedId: item.normalizedId,
             deps,
             file: item.id,
             variation: item.variation || this.config.baseConfig.dir,
